refactor(spells): split spellHelpers into individual modules
Each function now has its own file following one-export-per-file rule:
- hasSpellGrants.ts
- getGrantSpellSelectionRuleIds.ts
- removeSelectionsForRuleIds.ts
- removeConfirmedSchoolsForRuleIds.ts
- spellSelectionsComplete.ts
